Update facilities should only be for places
