+++ culturefeed_udb3.routing.yml
@@ -104,16 +104,6 @@ culturefeed_udb3.event.update_contactpoint:
     _format: json
     _method: POST
 
-culturefeed_udb3.event.update_facilities:
-  path: '/udb3/api/1.0/event/{cdbid}/facilities'
-  defaults:
-    _controller: '\Drupal\culturefeed_udb3\Controller\EventRestController::updateFacilities'
-    _title: 'Update facilities'
-  requirements:
-    _culturefeed_current_user: 'TRUE'
-    _format: json
-    _method: POST
-
 culturefeed_udb3.event_create:
   path: '/udb3/api/1.0/event'
   defaults:
